Applied simple patch as suggested by Doug Kosovic: http://www.freebsd.org/cgi/cvsweb.cgi/ports/devel/uclmmbase/files/patch-word Use uint32_t instead of `unsigned long', where a 32-bit unsigned value is needed -- this fixes the qfDES encryption/decryption on 64-bit platforms.

+++ src/qfDES.c
@@ -12,7 +12,7 @@ Added 2 August 1996, Saleem
 #include "qfDES.h"
 #include "crypt_random.h"
 
-typedef unsigned long Word;
+typedef uint32_t Word;
 
 #define B00 0x80000000
 #define B01 0x40000000
